refactor: ProposalQuestionListPane
