+++ packages/desktop/views/dashboard/governance/components/proposal-details/index.ts
@@ -1,3 +1,4 @@
 export { default as ProposalAccountVotingPane } from './ProposalAccountVotingPane.svelte'
 export { default as ProposalDetailsPane } from './ProposalDetailsPane.svelte'
 export { default as ProposalInformationPane } from './ProposalInformationPane.svelte'
+export { default as ProposalQuestionListPane } from './ProposalQuestionListPane.svelte'
